requirements.txt
Flask
tensorflow
numpy
Pillow
fpdf2
pydicom
opencv-python-headless
google.generativeai
pandas
scikit-learn
seaborn